separate swap logic from transfer logic
